homework/02/lib/lib.go
package lib

import (
	"encoding/json"
	"fmt"
	"os"
)

// SVar  marshals and indents a data structure into JSON.
func SVar(v interface{}) string {
	// s, err := json.Marshal ( v )
	s, err := json.Marshal(v)
	if err != nil {
		return fmt.Sprintf("Error:%s", err)
	}
	return string(s)
}

// SVarI  marshals and indents a data structure into JSON.
func SVarI(v interface{}) string {
	// s, err := json.Marshal ( v )
	s, err := json.MarshalIndent(v, "", "\t")
	if err != nil {
		return fmt.Sprintf("Error:%s", err)
	}
	return string(s)
}

// Exists returns true if a file or directory exists.
func Exists(name string) bool {
	if _, err := os.Stat(name); err != nil {
		if os.IsNotExist(err) {
			return false
		}
	}
	return true
}
